Fix API Order

+++ api/order/orderController.js
@@ -3,8 +3,27 @@ const http = require('../../const')
 const URLParser = require('../../helper/common/index')
 const productService = require('../products/productService')
 const mailService = require('../email/emailService')
+const storeService = require('../stores/storeService')
 exports.changeOrderStatus = async (req, res) => {
+
+    if (!req.body.store_id || !req.body.status){
+        res.status(http.BadRequest).json({
+            statusCode: http.BadRequest,
+            message: "Bad Request"
+        })
+        return
+    }
+
+    let check = {
+        id: req.body.store_id,
+        user_id: req.user.id
+    }
+    let authen = await this.AuthenticateUserAndStore(req, res, check)
+    if (!authen) {
+        return
+    }
     const query = req.body
+    delete query["store_id"]
     query.order_id = req.params.id
     const result = await orderService.changeOrderStatus(query)
 
@@ -82,6 +101,21 @@ exports.getOrder = async (req, res) => {
 }
 
 exports.deleteOrder = async (req, res) => {
+    if (!req.body.store_id){
+        res.status(http.BadRequest).json({
+            statusCode: http.BadRequest,
+            message: "Bad Request"
+        })
+        return
+    }
+    let check = {
+        id: req.body.store_id,
+        user_id: req.user.id
+    }
+    let authen = await this.AuthenticateUserAndStore(req, res, check)
+    if (!authen) {
+        return
+    }
     const orderId = req.params.id
     await orderService.deleteOrderStatus({order_id : orderId})
     await orderService.deleteOrderProducts({order_id : orderId})
@@ -90,7 +124,7 @@ exports.deleteOrder = async (req, res) => {
         res.status(http.Success).json({
             statusCode: http.Success,
             data: result,
-            message: "Successfully Change Status"
+            message: "Successfully Delete Order"
         })
     }
     else {
@@ -99,5 +133,26 @@ exports.deleteOrder = async (req, res) => {
             data: result,
             message: "No data found"
         })
+    }
+}
+
+exports.AuthenticateUserAndStore = async (req, res, check) => {
+    if (req.user) {
+        let query = {
+            id: check.id,
+            user_id: check.user_id
+        }
+        console.log(query)
+        const authenticateUser = await storeService.FindUserAndStore(query)
+        if (!authenticateUser[0]) {
+            res.status(http.Forbidden).json({
+                statusCode: 403,
+                message: "Forbiden!"
+            })
+            return
+        }
+        else {
+            return authenticateUser
+        }
     }
 }
